pubsubsac/requirements.txt
flask
requests
google-cloud-firestore
google-cloud-pubsub
gunicorn
